This project is my first attempt to integrate shared objects from SILGraphite smart font with a simple activity in Android Studio 6.0 with gradle plug in 0.11.
The shared objects are in jniLibs.  They are available only for armeabi and armeabi-v7a.  The Java wrapper for graphite c++ code is in org.sil.palaso.Graphite.  Calls to the Graphite Java object is placed in test.example.callgraphite.app.MainActivity.

Currently, the libraries loads but it is dying in the first call to add font.  It is getting a memory access exception:
06-17 11:25:09.950  13114-13114/test.example.callgraphite.app A/libc﹕ Fatal signal 11 (SIGSEGV) at 0x00000000 (code=1)